+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueEffectHandler.kt
@@ -70,6 +70,11 @@ class OverdueEffectHandler @AssistedInject constructor(
                 removedFromOverdueAppointments = overdueSections[Outcome.RemovedFromOverdueList].orEmpty(),
                 moreThanAnYearOverdueAppointments = overdueAppointmentMoreThanYear
             )
+
+
+
+
+            
           }
     }
   }