+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Show call result section by default in ContactPatientBottomSheet
 - Bump lottie to v5.2.0
 - Render overdue appointments when appointments are loaded and overdue sections feature is enabled
+- Add feature flag for overdue section improvements
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes