+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add feature flag for overdue section improvements
 - Load overdue appointments when current facility is loaded and overdue sections feature is enabled
 - Rename agreed to visit call result text
+- Change colour of number and chevron in the section headers to blue
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes