+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Change colour of number and chevron in the section headers to blue
 - Rename remind to call later call result text
 - Show all overdue patients count instead of just pending patients in overdue tab
+- Show progress bar when loading all overdue appointments
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes