+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Rename remind to call later call result text
 - Show all overdue patients count instead of just pending patients in overdue tab
 - Add query to load appointments for a patient
+- Added side effect for loading appointments for a patient
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes