+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Rename remind to call later call result text
 - Show all overdue patients count instead of just pending patients in overdue tab
 - Show progress bar when loading all overdue appointments
+- Implement minor overdue section layout improvements
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes